Add API debugging support with enhanced logging
This commit adds a logger to the provider to help with debugging API
calls. The logger is disabled by default and can be enabled by setting
the environment variable `MIST_API_DEBUG=true`. Test cases have been
added to verify that the provider is functioning as expected and the
refactoring that has been made does not break any existing functionality