fix deviceprofile_assign
fir for #70
in some condition the assigned/unassigned devices process may fail on the Mist side, and the list of MAC addresses returned may not match what is expected. This case was not processed by the provider, resulting a "This is a bug in the provider, which should be reported in the provider's own issue tracker." error message.
The provider is now returning a specific warning when it was not able to unassign a device, and a specific error when it was not able to assign a device